# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop-api Documentation",
        "description": "petshop api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088",
            "description": "local server"
        }
    ],
    "paths": {
...
        "/unidades": {
            "get": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Listar todas as Unidades",
                "description": "Retorna uma lista de todas as Unidades",
                "operationId": "03e31c9168ca621d946b2174c53dc3c5",
                "responses": {
                    "200": {
                        "description": "Lista de Unidades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Unidade"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Criar uma nova Unidade",
                "description": "Cria uma nova Unidade com base nos dados fornecidos",
                "operationId": "672e721c1125e57d1c5f9266972eef3f",
                "requestBody": {
                    "description": "Dados da nova Unidade",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Unidade"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Unidade criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unidade"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/unidades/{id}": {
            "get": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Obter detalhes de uma Unidade",
                "description": "Retorna os detalhes de uma Unidade com base no ID",
                "operationId": "e0a756cd9767913d8ca3125cf6d904c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Unidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da Unidade",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unidade"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Unidade não encontrada"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Atualizar uma Unidade",
                "description": "Atualiza uma Unidade com base no ID e nos dados fornecidos",
                "operationId": "c7b7443ada932707db67c7f8fbcf44b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Unidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados atualizados da Unidade",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Unidade"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unidade atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unidade"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Unidade não encontrada"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Excluir uma Unidade",
                "description": "Exclui uma Unidade com base no ID",
                "operationId": "776ae16045b8d9ac811011d9942be8dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Unidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Unidade excluída com sucesso"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Unidade não encontrada"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Unidade": {
                "title": "Unidade",
                "description": "Modelo de dados para uma Unidade",
                "properties": {
                    "id": {
                        "description": "ID da Unidade",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nome": {
                        "description": "Nome da Unidade",
                        "type": "string"
                    },
                    "endereco": {
                        "description": "Endereço da Unidade",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}